+++ src/stdune/list.mli
@@ -15,6 +15,8 @@ val filteri : 'a t -> f:(int -> 'a -> bool) -> 'a t
 
 val concat_map : 'a t -> f:('a -> 'b t) -> 'b t
 
+val rev_map_append : 'a t -> 'b t -> f:('a -> 'b) -> 'b t
+
 val     partition_map : 'a t -> f:('a -> ('b, 'c) Either.t) -> 'b t * 'c t
 val rev_partition_map : 'a t -> f:('a -> ('b, 'c) Either.t) -> 'b t * 'c t